Fix travis and started packages refactor
