+++ src/main/scala/geotrellis/pointcloud/store/avro/PointCloudCodecs.scala
@@ -1,33 +1,17 @@
-/*
- * Copyright 2017 Azavea
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package geotrellis.pointcloud.spark.io.avro.codecs
+package geotrellis.pointcloud.store.avro
 
 import geotrellis.store.avro._
 
-import io.pdal._
-import org.apache.avro._
-import org.apache.avro.generic._
+import io.pdal.{DimType, PointCloud, SizedDimType}
+import org.apache.avro.generic.GenericRecord
 import org.apache.avro.util.Utf8
+import org.apache.avro.{Schema, SchemaBuilder}
 import org.locationtech.jts.geom.Coordinate
 
-import java.util
-import java.nio.ByteBuffer
-
 import scala.collection.JavaConverters._
+
+import java.nio.ByteBuffer
+import java.util
 
 trait PointCloudCodecs extends Serializable {
   implicit def coordinateCodec = new AvroRecordCodec[Coordinate] {
