+++ src/main/scala/geotrellis/pointcloud/store/avro/package.scala
@@ -1,0 +1,3 @@
+package geotrellis.pointcloud.store
+
+package object avro extends Implicits
